Fixing and linting code
