--- app/storage/jsonl_store.py
@@ -1,14 +1,0 @@
-from __future__ import annotations
-import json
-from pathlib import Path
-from typing import Dict, Any
-
-
-class JSONLStore:
-    def __init__(self, path: Path) -> None:
-        self.path = path
-        self.path.parent.mkdir(parents=True, exist_ok=True)
-
-    def append(self, record: Dict[str, Any]) -> None:
-        with self.path.open("a", encoding="utf-8") as f:
-            f.write(json.dumps(record, ensure_ascii=False) + "\n")
